+++ integration-tests/package.json
@@ -50,8 +50,10 @@
 		"@aws-sdk/client-rds-data": "^3.549.0",
 		"@aws-sdk/credential-providers": "^3.549.0",
 		"@effect/sql-pg": "4.0.0-beta.58",
+		"@effect/sql-pglite": "4.0.0-beta.58",
 		"@effect/vitest": "4.0.0-beta.58",
-		"@electric-sql/pglite": "0.2.12",
+		"@electric-sql/pglite": "0.4.6",
+		"@electric-sql/pglite-postgis": "^0.0.8",
 		"@miniflare/d1": "^2.14.4",
 		"@miniflare/shared": "^2.14.4",
 		"@planetscale/database": "^1.16.0",
